multipathd/multipathd-shutdown-if-idle.service: New service for "shutdown if-idle"
On top of the new "shutdown if-idle" command, we use a
oneshot (actually "exec") systemd service that will be executed 120
seconds after booting up the system or after multipathd.service is
active, to send the command mentioned above to the daemon.

Note that this approach is still not perfect: if new devices are added
to the system while it's running, multipathd will be started, even if
no multipath'ing is done.  But that's better than the current state
where the daemon stays on unconditionally.

Signed-off-by: Sergio Durigan Junior <sergiodj@sergiodj.net>