multipathd: the local path change is not considered
When update_prio is called, there is a check to see if the local path
has a priority change. Then all the remaining paths are simliarly
checked.

Only the result of paths not matching the local path are
considered in the calculation of 'changed'. In the case of failed paths
becoming again available this can lead to multipath not reloading.

The result will look like this:
3624a93703c9f34490f6140a100011010 dm-2 PURE,FlashArray
size=3.0T features='1 retain_attached_hw_handler' hwhandler='1 alua' wp=rw
|-+- policy='service-time 0' prio=50 status=active
| |- 1:0:0:1  sdb 8:16  active ready running
| |- 9:0:0:1  sdc 8:32  active ready running
| |- 11:0:0:1 sdd 8:48  active ready running
| `- 13:0:0:1 sdh 8:112 active ready running
`-+- policy='service-time 0' prio=50 status=enabled
  |- 8:0:0:1  sde 8:64  active ready running
  |- 10:0:0:1 sdf 8:80  active ready running
  |- 12:0:0:1 sdg 8:96  active ready running
  `- 14:0:0:1 sdi 8:128 active ready running

Where the path's priorities are updated, but the path group is not
activated.

Fixes: 6ccd7b8 ("multipathd: only refresh priorities in update_prio()")
Signed-off-by: Brian Bunker <brian@purestorage.com>
Signed-off-by: Seamus Connor <sconnor@purestorage.com>
Signed-off-by: Krisha Kant <krishna.kant@purestorage.com>
Signed-off-by: Martin Wilck <mwilck@suse.com>
Reviewed-by: Benjamin Marzinski <bmarzins@redhat.com>