libmultipath: stop PURE FlashArray from detecting priority
If 'detect_prio' is set to 'yes', the 'sysfs' prioritizer will be used
instead of the 'alua' prioritizer. The 'sysfs' prioritizer can leave
paths in the Active/Non-optimized ALUA state after a successful
upgrade of a PURE FlashArray. Explicitly use the 'alua' prioritizer
instead for PURE FlashArray rather than the default of 'detect_prio'
which is 'yes'.

Signed-off-by: Brian Bunker <brian@purestorage.com>
Reviewed-by: Martin Wilck <mwilck@suse.com>