libmpathutil: really always use glibc basename()
Despite 03a5456 ("libmultipath: always use glibc basename()"), we
still call the system library's basename() from libmultipath.
musl libc until 1.24 provided a prototype for basename() in string.h,
which was not correct and was resolved to the destructive POSIX
basename(). musl libc 1.25 removed this prototype.

While the remaining code path doesn't strictly depend on the non-destructive
behavior of glibc's basename(), it's cleaner and safer to use the same
implementation everywhere.

Fixes: 03a5456 ("libmultipath: always use glibc basename()")
Fixes: https://github.com/opensvc/multipath-tools/pull/84

Signed-off-by: Martin Wilck <mwilck@suse.com>
Reviewed-by: Khem Raj <raj.khem@gmail.com>
Reviewed-by: Benjamin Marzinski <bmarzins@redhat.com>