multipathd: Set CPUWeight to 1000 and LimitRTPRIO to 10
If multipathd doesn't become a real time process, it was scheduled as a
normal process, without any priority increase. Bump up the CPUWeight so
that even as a normal process, it will still run with increased
priority.

If multipathd did become a real time process, it set itself to the
highest priority, which is excessive. A priority of 10 is plenty.

Signed-off-by: Benjamin Marzinski <bmarzins@redhat.com>
Reviewed-by: Martin Wilck <mwilck@suse.com>