libmultipath: accept poorly chosen aliases in find_mp_by_str
A multipath device alias of dm-1 is valid, albeit a really bad idea.
Make sure that find_mp_by_str() is only checking for multipath devices
by minor number for strings that really are of the form dm-<number>, and
if the minor-number check fails, check if the string is an alias, just
to be safe.

Signed-off-by: Benjamin Marzinski <bmarzins@redhat.com>
Reviewed-by: Martin Wilck <mwilck@suse.com>