multipathd: handle changed wwid when adding partial path
If multipathd noticed that the WWID didn't match the device when adding
a partial path, but removing it failed, multipathd wasn't disabling the
path. Instead of calling handle_path_wwid_change(), which doesn't make
much sense when multipathd is expecting a uevent, just manually disable
the path if ev_remove_path() fails.

Since the path hasn't actually been removed, multipathd needs to make
sure that it says failed when the path state is resynced with the
kernel. Set the state to PATH_UNCHECKED in orphan_path(), and
additionally skip all path in INIT_REMOVED in sync_map_state().

Signed-off-by: Benjamin Marzinski <bmarzins@redhat.com>
Reviewed-by: Martin Wilck <mwilck@suse.com>