multipathd: show pending reconfigs in 'show daemon' output
Since the "multipathd reconfigure" command simply schedules a
reconfigure, and doesn't wait until it has occurred, it would be nice to
be able to query multipathd to see if the reconfigure has completed.
"multipathd show daemon" doesn't work for this. The reconfigure could be
delayed, and the output gives no indication that there is a pending
reconfigure.

Change the ouput to add " (pending reconfigure)" if a reconfigure is
pending.

Signed-off-by: Benjamin Marzinski <bmarzins@redhat.com>
Reviewed-by: Martin Wilck <mwilck@suse.com>