libmpathutil.version: remove strlcpy and strlcat, and LIBMULTIPATH section
With the removal of strlcpy from the "symbols referenced by multipath
and multipathd" section, this part of the ABI (which is actually the
part of LIBMULTIPATH_16.0.0 which was moved to libmpathutil and exported
to multipath and/or multipathd) has now changed for the first time since
the split between libmpathutil and libmultipath.

It makes no sense to further track whether or not any symbols exported
by libmpathutil are used by multipath or multipathd. It turns out that
both executables were already using some symbols from the sections that
were supposedly used by libmultipath only.

Therefore, move all symbols to the new LIBMPATHUTIL_4.0 ABI.

Signed-off-by: Martin Wilck <mwilck@suse.com>
Reviewed-by: Benjamin Marzinski <bmarzins@redhat.com>