multipath-tools: use common convention for "header file included" macros
User code shouldn't define any identifiers starting with 2 underscores, or
underscore + capital letter [1]. Adopt this convention for all macros
indicating that a certain header file is included. Moreover, use a common
convention for constructing these macros: the macro is the base name of the
header file, in capital letters, non-alphanumeric characters replaced by
underscore, with "_INCLUDED" appended. This convention is well readable and
was already used in a few files.

The only exceptions to this rule are:
- the header files under libmultipath/nvme and libmultipath/nvme/linux, where
"NVME_" or "LINUX_" is prepended to the base name, respectively;
- the files "byteorder.h" and "devmapper.h" under kpartx/, to avoid ambiguity
with the files by the same name under libmultipath;
- "main.h" for mpathpersist, to avoid ambiguity with multipathd's main.h.

This commit also adds "already included" macros where they'd been missing
so far.

[1] https://www.gnu.org/software/libc/manual/html_node/Reserved-Names.html

Signed-off-by: Martin Wilck <mwilck@suse.com>
Reviewed-by: Benjamin Marzinski <bmarzins@redhat.com>