libmpathutil: avoid extra memory allocation in print_strbuf()
Instead of allocating a temporary buffer with vasprintf() and copying its
contents to the strbuf, try to print directly into the strbuf and use the
return value to expand the buffer if necessary. This saves a memory
allocation on every invocation of print_strbuf(), and thus reduces the
likelihood of failure and at the same time increases efficiency. We need
an extra call to vsnprintf() if the buffer must be expanded, but vasprintf()
will need to re-process the input internally as well after determining the
required buffer size, therefore this probably won't hurt much.

Add some more tests for print_strbuf() to make sure this works as intended.

Signed-off-by: Martin Wilck <mwilck@suse.com>
Reviewed-by: Benjamin Marzinski <bmarzins@redhat.com>