libmultipath: reduce log level of "map X has multiple targets"
On systems with LVM volumes, "multipath -ll" will spit out lots of messages
like

    libmp_mapinfo: map vg-lv0 has multiple targets

which is irritating. Reduce the log level of these messages to 3, as they
are harmless most of the time.

This is a backport of e8949c2 ("libmultipath: reduce log level of
libmp_mapinfo() messages") from the master branch. We can't apply exactly
the same fix because the stable branch is missing 8c772d3 ("libmultipath:
check DM UUID earlier in libmp_mapinfo__").

Signed-off-by: Martin Wilck <mwilck@suse.com>