multipathd: set rport port_state to marginal for NVMe devices
When a scsi path device is set to marginal, it updates the rport state.
Do this for NVMe devices as well. Unlike with scsi devices, setting the
rport state to marginal will not directly change the nvme driver
behavior. However it will allow users to see the rports associated with
their Marginal paths.

Signed-off-by: Benjamin Marzinski <bmarzins@redhat.com>
Acked-by: Muneendra Kumar <muneendra.kumar@broadcom.com>
Reviewed-by: Martin Wilck <mwilck@suse.com>