libmultipath: don't set need_reload in adopt_paths
adopt_paths() is generally called when you are planning on creating or
reloading a multipath device. The only time when adopt_paths() is called
and mpp->action isn't explicitly set to ACT_CREATE or ACT_RELOAD
afterwards is when it is called by coalesce_paths() ->
add_map_with_path(). But in this case, you will only set need_reload if
the path that is being adopted isn't already part of the device. Because
of that, the new device that's being assembled in coalesce_paths() must
be different from the existing device (it has a path that isn't in the
existing device), so select_action() will select ACT_RELOAD, regardless
of the need_reload state. This makes intuitive sense. When multipath
adopts a new path, it should reload the device to actually add that path
to the device.

Signed-off-by: Benjamin Marzinski <bmarzins@redhat.com>
Reviewed-by: Martin Wilck <mwilck@suse.com>